refactor(posts): move meeting tab to its own module
