+++ features/posts/aggregate/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/aggregate/presentation/di/FeaturesPostPresentationModule.kt
@@ -1,9 +1,10 @@
 package com.parksupark.soomjae.features.posts.aggregate.presentation.di
 
 import com.parksupark.soomjae.features.posts.aggregate.presentation.post.PostViewModel
-import com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.meeting.MeetingTabViewModel
 import com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.member.MemberTabViewModel
 import com.parksupark.soomjae.features.posts.common.presentation.di.featuresPostsCommonPresentationModule
+import com.parksupark.soomjae.features.posts.common.presentation.di.featuresPostsMeetingPresentationModule
+import com.parksupark.soomjae.features.posts.common.presentation.tab.MeetingTabViewModel
 import com.parksupark.soomjae.features.posts.community.presentation.di.featuresPostsCommunityPresentationModule
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
@@ -17,5 +18,5 @@ private val postModule = module {
 
 val featuresPostsAggregatePresentationModule = module {
     includes(postModule)
-    includes(featuresPostsCommonPresentationModule, featuresPostsCommunityPresentationModule)
+    includes(featuresPostsCommonPresentationModule, featuresPostsCommunityPresentationModule, featuresPostsMeetingPresentationModule)
 }
