+++ features/posts/meeting/presentation/src/androidMain/kotlin/com/parksupark/soomjae/features/posts/common/presentation/previews/MeetingTab.kt
@@ -1,10 +1,10 @@
-package com.parksupark.soomjae.features.posts.aggregate.presentation.previews
+package com.parksupark.soomjae.features.posts.common.presentation.previews
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.parksupark.soomjae.core.presentation.designsystem.theme.AppTheme
-import com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.meeting.MeetingTabScreen
-import com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.meeting.MeetingTabState
+import com.parksupark.soomjae.features.posts.common.presentation.tab.MeetingTabScreen
+import com.parksupark.soomjae.features.posts.common.presentation.tab.MeetingTabState
 
 @Composable
 @Preview(name = "MeetingTab")
@@ -12,7 +12,7 @@ private fun MeetingTabScreenPreview() {
     AppTheme {
         MeetingTabScreen(
             state = MeetingTabState(),
-            onAction = {},
+            onAction = { },
         )
     }
 }
