+++ features/posts/meeting/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/presentation/tab/MeetingTabScreen.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.meeting
+package com.parksupark.soomjae.features.posts.common.presentation.tab
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
