+++ features/posts/meeting/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/presentation/tab/MeetingTabViewModel.kt
@@ -1,14 +1,11 @@
-package com.parksupark.soomjae.features.posts.aggregate.presentation.post.tabs.meeting
+package com.parksupark.soomjae.features.posts.common.presentation.tab
 
-import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 
-class MeetingTabViewModel(
-    savedStateHandle: SavedStateHandle,
-) : ViewModel() {
+class MeetingTabViewModel : ViewModel() {
     private val _stateFlow: MutableStateFlow<MeetingTabState> = MutableStateFlow(MeetingTabState())
     val stateFlow: StateFlow<MeetingTabState> = _stateFlow.asStateFlow()
 }
